Minor changes that might not be necessary in the future
